backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pgvector==0.2.3

# Redis and Celery
redis==5.0.1
celery==5.3.4
flower==2.0.1

# AI/LLM
openai==1.3.7
anthropic==0.7.1
tiktoken==0.5.1

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
WeasyPrint==60.1
pypdf==3.17.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3

# Email
sendgrid==6.11.0

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
psutil==5.9.6

# Security
cryptography==41.0.7
bcrypt==4.1.2
python-multipart==0.0.6

# Performance & Caching
redis==5.0.1
aioredis==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Load Testing
locust==2.17.0

